+++ apps/web/components/Basics.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { AutoForm as ShadcnRHFAutoForm } from "@dual-autoform/shadcn/components/ui/autoform/react-hook-form";
 import { AutoForm as ShadcnTanstackAutoForm } from "@dual-autoform/shadcn/components/ui/autoform/tanstack-form";
-import { AutoFormFieldProps } from "@dual-autoform/react";
+import { AutoFormFieldProps } from "@dual-autoform/react/react-hook-form";
 import { zodSchemaProvider } from "./utils";
 import "@dual-autoform/shadcn/globals.css";
 
@@ -36,9 +36,6 @@ export function BasicsTanstack() {
   return (
     <ShadcnTanstackAutoForm
       schema={zodSchemaProvider}
-      onFormInit={(f) => {
-        (window as any).form = f;
-      }}
       onSubmit={(data) => {
         console.log(JSON.stringify(data, null, 2));
       }}