+++ packages/mantine/src/index.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-export {
-  MantineAutoFormFieldComponents,
-  MantineUIComponents,
-  createAutoForm,
-} from "./AutoForm";
-export type { FieldTypes } from "./AutoForm";
-export * from "./types";