+++ packages/mantine/tsup.config.ts
@@ -2,7 +2,6 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   entry: {
-    index: "src/index.tsx",
     "react-hook-form": "src/react-hook-form.tsx",
     "tanstack-form": "src/tanstack-form.tsx",
   },