+++ packages/mui/src/index.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-export {
-  MuiAutoFormFieldComponents,
-  MuiUIComponents,
-  createAutoForm,
-} from "./AutoForm";
-export type { FieldTypes } from "./AutoForm";
-export * from "./types";