+++ packages/react/tsup.config.ts
@@ -3,8 +3,8 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   entry: {
     index: "src/index.tsx",
-    "react-hook-form/index": "src/react-hook-form/index.ts",
-    "tanstack-form/index": "src/tanstack-form/index.ts",
+    "react-hook-form": "src/react-hook-form/index.ts",
+    "tanstack-form": "src/tanstack-form/index.ts",
   },
   format: ["esm", "cjs"],
   outExtension({ format }) {
@@ -13,5 +13,4 @@ export default defineConfig({
   dts: true,
   clean: true,
   sourcemap: true,
-  external: ["@dual-autoform/react"],
 });